# yamls/swagger.yaml
swagger: '2.0'
info:
  title: memecoins App api
  description: memecoins App api swagger文档
  termsOfService: ''
  contact:
    email: sales@techrare.com
  version: 1.0.0
host: office.techrare.com
basePath: /memecoinsapi/public/api
schemes:
  - https
consumes:
  - application/x-www-form-urlencoded
paths:
...
  /upload_base64:
    post:
      consumes:
        - multipart/form-data
      tags:
        - tools
      summary: 上传base64图片
      description: 上传base64图片
      operationId: upload_base64
      parameters:
        - name: file
          in: formData
          description: 圖片
          required: true
          type: string
        - name: dir
          in: formData
          description: 存放目錄
          enum:
            - avatar
            - banner
            - activity
            - temp
          default: temp
          required: true
          type: string
      responses:
        '200':
          description: '圖片上傳成功,并返回图片路徑'
        '400':
          description: 沒有圖片上傳
      security:
        - Bearer: []
...
  /share:
    get:
      tags:
        - tools
      summary: 分享
      description: 分享
      operationId: share
      parameters:
        - name: id
          in: query
          description: 店鋪id
          required: true
          type: integer
      responses:
        '200':
          description: >-
            生成頁面成功,https://office.techrare.com/memecoinsapi/public/share/9999.html
        '401':
          description: 頁面生成失敗
      security:
        - Bearer: []
...
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header